add opencompass_evaluator.py
